solve error with wrong coordinate, because of wrong type of data used in organize.c

+++ fillit.h
@@ -6,7 +6,7 @@
 /*   By: kda-fons <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/01/09 12:50:26 by kda-fons          #+#    #+#             */
-/*   Updated: 2017/01/16 23:20:30 by psebasti         ###   ########.fr       */
+/*   Updated: 2017/01/17 16:54:12 by psebasti         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ int			check_tetri(char *str);
 int			check_valid(char *str);
 int			check_four_blocks(char *str);
 size_t		organize(t_tetri *tet);
-size_t		tet_min_xy(t_tetri *tet, int xy);
+size_t		tet_min_xy(t_tetri *tet, size_t xy);
 void		mv_upleft(t_tetri *tet, size_t *min);
 size_t		calc_min_square(size_t numtetri);
 size_t		solve(t_tetri *tet, size_t numtetri);
